+++ src/__tests__/data.test.ts
@@ -1,6 +1,6 @@
 import { expect } from "@std/expect";
 
-import { normalizeString } from "../helpers/data.ts";
+import { normalizeString } from "../utils/data.ts";
 
 Deno.test("normalizeString should return a string with no leading or trailing whitespace", () => {
   const actual = normalizeString("  hello  ");
