+++ src/db/queries/index.ts
@@ -1,0 +1,2 @@
+export * from "./geolocation.ts";
+export * from "./municipalities.ts";
