+++ src/db/types/index.ts
@@ -1,0 +1,2 @@
+export * from "./geolocation-schema.ts";
+export * from "./municipalities-schema.ts";
