+++ src/db/types/municipalities-schema.ts
@@ -1,0 +1,5 @@
+export type Municipality = {
+  id: number;
+  municipality: string;
+  loan: string;
+};
