--- src/helpers/db/index.ts
@@ -1,1 +1,0 @@
-export * from "./client.ts";
