--- src/lib/index.ts
@@ -1,2 +1,0 @@
-export * from "./municipalities.ts";
-export * from "./geolocation.ts";
