+++ src/services/nominatim.ts
@@ -21,3 +21,16 @@ export type NominatimData = {
   };
   boundingbox: string[];
 };
+
+export const getNominatimGeolocationByCity = async (
+  municipality: string,
+): Promise<NominatimData[]> => {
+  console.log(`Fetching data for ${municipality}`);
+
+  const data = await fetch(
+    `https://nominatim.openstreetmap.org/search?format=json&addressdetails=1&city=${municipality}&country=nl&limit=1`,
+  );
+  const res = await data.json();
+
+  return res;
+};
