--- src/types/api/index.ts
@@ -1,2 +1,0 @@
-export * from "./postgresql.ts";
-export * from "./nominatim.ts";
