+++ src/utils/csv/merge-data.ts
@@ -1,10 +1,10 @@
-import { normalizeString } from "../helpers/index.ts";
+import { normalizeString } from "../index.ts";
 import type {
   MergeDataResult,
   MergedDataEntry,
   ParsedCbsDataEntry,
   ParsedSvnDataEntry,
-} from "../types/index.ts";
+} from "../../types/index.ts";
 
 /**
  * Merges SVN data entries with CBS data entries based on matching municipalities.
